pyproject.toml
[tool.poetry]
name = "english"
version = "0.1.0"
description = ""
authors = ["PranitSh <ppshah2023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
inform = "^1.27.0"
clipboard = "^0.0.4"
pathlib = "^1.0.1"
pyyaml = "^6.0"
nltk = "^3.8.1"
regex = "^2022.10.31"
pytz = "^2022.7.1"
openai = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
